Create an interface for each type of move action
