+++ frontend/src/components/controls/new/NewGame.tsx
