+++ frontend/src/services/tempData.json
@@ -1,15 +1,15 @@
 {
   "sudoku": {
     "problem": [
-      [9, null, null, 5, null, null, null, null, null],
-      [6, 8, 7, 3, 4, null, null, 5, null],
-      [null, null, 4, 1, null, 7, null, 3, null],
-      [null, 4, null, null, 2, 6, null, null, 8],
-      [null, 2, 6, null, 5, 1, 4, 9, null],
-      [null, 9, 1, 4, null, 3, null, 6, 2],
-      [null, null, 2, 7, null, null, 9, null, 5],
-      [null, null, null, null, null, null, 3, null, 1],
-      [5, null, null, null, null, 9, 6, null, 7]
+      [9, null, null, 5, 6, 8, 2, 7, 4],
+      [6, 8, 7, 3, 4, 2, 1, 5, 9],
+      [2, 5, 4, 1, 9, 7, 8, 3, 6],
+      [3, 4, 5, 9, 2, 6, 7, 1, 8],
+      [7, 2, 6, 8, 5, 1, 4, 9, 3],
+      [8, 9, 1, 4, 7, 3, 5, 6, 2],
+      [1, 6, 2, 7, 3, 4, 9, 8, 5],
+      [4, 7, 9, 6, 8, 5, 3, 2, 1],
+      [5, 3, 8, 2, 1, 9, 6, 4, 7]
     ],
     "solution": [
       [9, 1, 3, 5, 6, 8, 2, 7, 4],
